+++ packages/components/src/ProgressCycle/index.ts
@@ -1,0 +1,1 @@
+export * from './ProgressCycle';
